Create dedicated Content Service for user-generated content
Architecture Refactoring:
- Extract recipe storage from Apps Service to dedicated Content Service
- Implement proper domain-driven design with service separation
- Content Service handles all app-specific user content storage

Content Service (Port 8006):
- FastAPI service with comprehensive recipe management API
- JWT authentication with user-scoped access control
- 10MB content limits with intelligent title extraction
- Mobile-first sync capabilities with conflict resolution
- Dedicated Dockerfile and Railway deployment configuration

Database Schema (Unchanged):
- user_recipes table with UUID, app_id, content, JSONB metadata
- Optimized indexing for user_id, app_id, favorites, and timestamps
- Support for multiple apps via app_id field

API Endpoints:
- GET /recipes/users/{user_id}/recipes - Paginated retrieval
- POST /recipes/users/{user_id}/recipes - Save new recipes
- PUT /recipes/users/{user_id}/recipes/{recipe_id} - Update favorites
- DELETE /recipes/users/{user_id}/recipes/{recipe_id} - Delete recipes
- POST /recipes/users/{user_id}/recipes/sync - Mobile bulk sync

Apps Service Cleanup:
- Removed all recipe-related models, routes, and imports
- Restored focus to app lifecycle management and LLM configuration
- Fixed railway.json merge conflict

Frontend Integration:
- New base URL: https://content.fairydust.fun/recipes
- All existing API contracts maintained
- No breaking changes to request/response formats

Benefits:
- Clean separation of concerns between app management and content storage
- Independent scaling for content vs app management workloads
- Extensible architecture for future app content types
- Proper microservices domain boundaries
