+++ services/content/railway.json
@@ -2,10 +2,10 @@
   "$schema": "https://railway.app/railway.schema.json",
   "build": {
     "builder": "DOCKERFILE", 
-    "dockerfilePath": "services/apps/Dockerfile"
+    "dockerfilePath": "services/content/Dockerfile"
   },
   "deploy": {
-    "startCommand": "cd services/apps && python main.py",
+    "startCommand": "cd services/content && python main.py",
     "numReplicas": 1,
     "healthcheckPath": "/health",
     "healthcheckTimeout": 30,
