+++ services/content/requirements.txt
@@ -1,0 +1,8 @@
+fastapi==0.104.1
+uvicorn==0.24.0
+pydantic==2.5.0
+asyncpg==0.29.0
+redis==5.0.1
+python-multipart==0.0.6
+python-jose[cryptography]==3.3.0
+python-dotenv==1.0.0
